DeepSeek-V4-Flash: quantitative evidence that initialize() never transfers weights
Stopped guessing and measured. Comparing our initialize() against the working
Qwen3 control's weight load:

  Qwen3-1.7B (works)     ~0.42 GB/rank,  8 ranks,  12.0 s
  DSV4 5 layers (zeros)  ~2.06 GB/rank, 32 ranks,   5.2 s

We hand over 4.9x the data per rank and finish in 43% of the time -- an ~11x
apparent transfer-rate difference. With neuron-monitor reporting zero runtimes on
the subsequent call, that says initialize() returns without moving weights to the
device.

This is narrower than "the graph does not execute" and it is testable, because we
also know initialize() is not simply ignoring the dict: it validates shapes (a
wrong-shape dict from shard_checkpoint raises "Incorrect tensor shape at
inner.embed.weight: received 129280 4096, expected 4040 4096") and all 207 graph
weight names reconcile with 0 missing. So the weights are right, the call inspects
them, and nothing lands on device.

Also tested and ruled out: int32 vs int64 input_ids, matching NxDI's own input
generator (model_wrapper.py:245). No change.

+++ contrib/models/DeepSeek-V4-Flash/src/run_dsv4_device.py
@@ -180,7 +180,7 @@ def _example_inputs(mode):
     #   status=1006 Execution Out-Of-Bounds Memory Access
     # Spread the example ids across the vocab so every rank traces the general
     # case. The values do not matter otherwise; only the shapes are recorded.
-    ids = (torch.arange(n_active, dtype=torch.long) * 977 + 101) % 129280
+    ids = (torch.arange(n_active, dtype=torch.int32) * 977 + 101) % 129280
     ids = ids.unsqueeze(0).repeat(_BATCH, 1)
     if mode == "decode":
         pos = torch.ones(_BATCH, 1, dtype=torch.int32)
@@ -361,7 +361,7 @@ def main():
         print(f"[dev] could not query is_initialized: {type(e).__name__}: {e}")
 
     def call(token_ids, positions):
-        inp = torch.tensor([token_ids], dtype=torch.long)
+        inp = torch.tensor([token_ids], dtype=torch.int32)
         pos = torch.tensor([positions], dtype=torch.int32)
         # Call nxd_model directly, NOT the traced NxDModelExecutor wrapper.
         # `traced` came out of torch.jit.trace, which recorded the ops of ONE
